Minor fixes and adjustments
made small changes in various files to improve code readability and consistency, no functional changes.
