+++ src/main/java/restservice/models/Genre.java
@@ -1,6 +1,5 @@
 package restservice.models;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 
 public enum Genre {
     Action,
